+++ app/src/main/java/org/simple/clinic/drugs/selection/custom/DrugInfoProgressState.kt
@@ -1,0 +1,6 @@
+package org.simple.clinic.drugs.selection.custom
+
+enum class DrugInfoProgressState {
+  IN_PROGRESS,
+  DONE
+}